fix: Collectable renamed to CollectionItem
