+++ Scripts/Editor/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e854de029fd501643b53b644882e0d3d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
