+++ Scripts/Editor/Core/CollectableItemIndirectReferenceDrawer.cs
@@ -3,47 +3,47 @@ using UnityEngine;
 
 namespace BrunoMikoski.ScriptableObjectCollections
 {
-    [CustomPropertyDrawer(typeof(CollectableIndirectReference), true)]
-    public sealed class CollectableIndirectReferenceDrawer : PropertyDrawer
+    [CustomPropertyDrawer(typeof(CollectionItemIndirectReference), true)]
+    public sealed class CollectableItemIndirectReferenceDrawer : PropertyDrawer
     {
         private const string OBJECT_ASSET_PROPERTY_PATH = "editorAsset";
-        private const string COLLECTABLE_GUID_PROPERTY_PATH = "collectableGUID";
+        private const string COLLECTION_ITEM_GUID_PROPERTY_PATH = "collectionItemGUID";
         private const string COLLECTION_GUID_PROPERTY_PATh = "collectionGUID";
 
 
         private SerializedProperty objectAssetProperty;
-        private SerializedProperty collectableGUIDProperty;
+        private SerializedProperty collectionItemGUIDProperty;
         private SerializedProperty collectionGUIDProperty;
 
-        private CollectableScriptableObject cachedReference;
+        private ScriptableObjectCollectionItem cachedReference;
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
             objectAssetProperty = property.FindPropertyRelative(OBJECT_ASSET_PROPERTY_PATH);
-            collectableGUIDProperty = property.FindPropertyRelative(COLLECTABLE_GUID_PROPERTY_PATH);
+            collectionItemGUIDProperty = property.FindPropertyRelative(COLLECTION_ITEM_GUID_PROPERTY_PATH);
             collectionGUIDProperty = property.FindPropertyRelative(COLLECTION_GUID_PROPERTY_PATh);
 
             if (objectAssetProperty.objectReferenceValue != null)
             {
-                if (string.IsNullOrEmpty(collectableGUIDProperty.stringValue)
+                if (string.IsNullOrEmpty(collectionItemGUIDProperty.stringValue)
                     || string.IsNullOrEmpty(collectionGUIDProperty.stringValue))
                 {
-                    CollectableScriptableObject collectable = objectAssetProperty.objectReferenceValue as CollectableScriptableObject;
-                    collectableGUIDProperty.stringValue = collectable.GUID;
+                    ScriptableObjectCollectionItem collectable = objectAssetProperty.objectReferenceValue as ScriptableObjectCollectionItem;
+                    collectionItemGUIDProperty.stringValue = collectable.GUID;
                     collectionGUIDProperty.stringValue = collectable.Collection.GUID;
                     objectAssetProperty.serializedObject.ApplyModifiedProperties();
                 }
             }
             else
             {
-                if (!string.IsNullOrEmpty(collectableGUIDProperty.stringValue)
+                if (!string.IsNullOrEmpty(collectionItemGUIDProperty.stringValue)
                     && !string.IsNullOrEmpty(collectionGUIDProperty.stringValue))
                 {
                     if (CollectionsRegistry.Instance.TryGetCollectionByGUID(collectionGUIDProperty.stringValue,
                         out ScriptableObjectCollection collection))
                     {
-                        if (collection.TryGetCollectableByGUID(collectableGUIDProperty.stringValue,
-                            out CollectableScriptableObject collectable))
+                        if (collection.TryGetCollectableByGUID(collectionItemGUIDProperty.stringValue,
+                            out ScriptableObjectCollectionItem collectable))
                         {
                             objectAssetProperty.objectReferenceValue = collectable;
                             objectAssetProperty.serializedObject.ApplyModifiedProperties();
@@ -53,7 +53,7 @@ namespace BrunoMikoski.ScriptableObjectCollections
             }
 
             if (cachedReference == null && objectAssetProperty.objectReferenceValue != null)
-                cachedReference = objectAssetProperty.objectReferenceValue as CollectableScriptableObject;
+                cachedReference = objectAssetProperty.objectReferenceValue as ScriptableObjectCollectionItem;
             
             EditorGUI.PropertyField(position, objectAssetProperty, label, true);
             if (objectAssetProperty.objectReferenceValue != cachedReference)
@@ -62,7 +62,7 @@ namespace BrunoMikoski.ScriptableObjectCollections
                 string collectionGUID = string.Empty;
             
                 if (objectAssetProperty.objectReferenceValue != null &&
-                    objectAssetProperty.objectReferenceValue is CollectableScriptableObject collectable)
+                    objectAssetProperty.objectReferenceValue is ScriptableObjectCollectionItem collectable)
                 {
                     collectableGUID = collectable.GUID;
                     collectionGUID = collectable.Collection.GUID;
@@ -73,7 +73,7 @@ namespace BrunoMikoski.ScriptableObjectCollections
                     cachedReference = null;
                 }
             
-                collectableGUIDProperty.stringValue = collectableGUID;
+                collectionItemGUIDProperty.stringValue = collectableGUID;
                 collectionGUIDProperty.stringValue = collectionGUID;
                 objectAssetProperty.serializedObject.ApplyModifiedProperties();
                 EditorUtility.SetDirty(objectAssetProperty.serializedObject.targetObject);
