+++ Scripts/Editor/Core/CollectableItemIndirectReferenceDrawer.cs.meta
