+++ Scripts/Editor/Core/CollectableItemObjectPropertyDrawer.cs.meta
