+++ Scripts/Editor/Core/CollectionPreprocessBuild.cs
