+++ Scripts/Editor/Core/CollectionsAssetsPostProcessor.cs.meta
