+++ Scripts/Editor/Core/CopyCollectableUtils.cs.meta
