+++ Scripts/Editor/Wizzard.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e1d20e0e0c4955a46938914ffc60b24f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
