+++ Scripts/Editor/Wizzard/CreateCollectionWizzard.cs.meta
