+++ Scripts/Runtime/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a8b09dda7014a3bb8a5818444450b3c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
