+++ Scripts/Runtime/Core/CollectableIndirectReference.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 0fd0d9dfccc34739af6081bb29ee470a
+timeCreated: 1617048538
