+++ Scripts/Runtime/Core/CollectableScriptableObject.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace BrunoMikoski.ScriptableObjectCollections
+{
+    [Obsolete("CollectableScriptableObject have been renamed to ScriptableObjectCollectionItem")]
+    public class CollectableScriptableObject : ScriptableObjectCollectionItem
+    {
+        
+    }
+}
