+++ Scripts/Runtime/Core/CollectableScriptableObject.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e0a16cd2a9d1422591f9e2695c90680e
+timeCreated: 1617048465
