+++ Scripts/Runtime/Core/CollectionItemIndirectReference.cs.meta
