+++ Scripts/Runtime/Core/CollectionsRegistry.cs.meta
