+++ Scripts/Runtime/Core/InitializeCollections.cs
