+++ Scripts/Runtime/Core/ResourceScriptableObjectSingleton.cs
