+++ Scripts/Runtime/Core/ResourceScriptableObjectSingleton.cs.meta
