+++ Scripts/Runtime/Core/ScriptableObjectCollection.cs.meta
