+++ Scripts/Runtime/Core/ScriptableObjectCollectionItem.cs.meta
