+++ Scripts/Runtime/Core/ScriptableObjectCollectionSettings.cs.meta
